build(cmake): allow import as subproject
CMake functions inside `cmake/third_party_deps.cmake` are converted to
macros, because macros are handled as text replacement, which leads to
the fact that the calls are executed in the scope of the script using
the macro, whereas functions create their own scope.
This is especially important for the pico-sdk, due to the fact that
their initialization won't work properly otherwise.