feat(spi): add alternative SPI implementation
This new implementation allows for easier setup and provides an interface that is completely independent from pico sdk.
Users only need to call `PicoSpi_create` and obtain an ADT that will handle all subsequent spi interaction.
Configuration of the spi interface is handled separately and not during instantiation.
This way configuration remains independent of the pico sdk as well.